Move DirectedGraph into a separate file
